Symbol class should contain just basic information
