+++ src/main/java/com/github/oscerd/finnhub/client/FinnhubClient.java
@@ -31,7 +31,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.oscerd.finnhub.model.CompanyProfile;
 import com.github.oscerd.finnhub.model.Exchange;
 import com.github.oscerd.finnhub.model.Quote;
-import com.github.oscerd.finnhub.model.Symbol;
+import com.github.oscerd.finnhub.model.EnrichedSymbol;
 import com.github.oscerd.finnhub.model.SymbolLookup;
 
 public class FinnhubClient {
@@ -96,7 +96,7 @@ public class FinnhubClient {
 		return objectMapper.readValue(result, CompanyProfile.class);
 	}
 	
-	public List<Symbol> getSymbols(String exchange) throws ClientProtocolException, IOException {
+	public List<EnrichedSymbol> getSymbols(String exchange) throws ClientProtocolException, IOException {
 		HttpGet get = new HttpGet(Endpoint.SYMBOL.url() + "?token=" + token + "&exchange=" + Exchange.valueOf(exchange).code());
 
 		String result = null;
@@ -104,7 +104,7 @@ public class FinnhubClient {
 			result = EntityUtils.toString(response.getEntity());
 		}
 
-		return objectMapper.readValue(result, new TypeReference<List<Symbol>>(){});
+		return objectMapper.readValue(result, new TypeReference<List<EnrichedSymbol>>(){});
 	}
 
 	public SymbolLookup searchSymbol(String query) throws ClientProtocolException, IOException {
